Multicorrect test triggers a BUG. If the main thread exits when another thread are writing, something bad happens. This may cause the branches to be inconsistent. Every branch write will succeed. But none of them write the exact length that the write operation wants to. So none of the branche write is MTFS_OPERATION_PREFERABLE but all of them are MTFS_OPERATION_SUCCESS. Actually, The first branch succeeds to write a N*4096 length, e.g. 91738112, but less than 104867600 which is the length of the operation wants to write. And the following branches(or branch) only succeed to write 4096 length. The generic_perform_write() explains why the problem happens. Every time it writes a page, it will call fatal_signal_pending() to check whether there is any signal pending. The write operations are interrupted by the signal, so return value of them are shorter than expected.
This patch handles different write length between branches. When write
lengthes are differennt, mtfs_oplist_gather_writev() will take the longest
write length as a success, and take others as failures. Dirty flags will
be set to these branches.

Multicorrt test are changed in this patch to skip diff two branches when
the dirty flag is set.

+++ mtfs/tests/raid.sh
@@ -223,8 +223,14 @@ test_4() {
 		FILE_LIST="$FILE_LIST $BRANCH_DIR/$DIR_SUB/$tfile"
 	done;
 
+	rm $DIR/$tfile -f
 	$MULTICORRECT $DIR/$tfile $FILE_LIST -s 60 > /dev/null \
 	|| error "multicorret failed"
+	$UTIL_MTFS getstate $DIR/$tfile | grep 0xa0 > /dev/null
+	if [ $? -eq 0 ]; then
+		echo "inconsistence set, skip diff"
+		return 0;
+	fi
 	diff $FILE_LIST > /dev/null || error "file diff"
 	rm -f $DIR/$tfile
 }
